refactor proposalTypes component, add test, stories, include translation setup in storybook
