+++ src/Components/ProposalTypes/components/ProposalTypeDescription/index.ts
@@ -1,0 +1,1 @@
+export { default as ProposalTypeDescription } from './ProposalTypeDescription';
