+++ src/Components/ProposalTypes/fixtures.ts
@@ -1,0 +1,18 @@
+import { ProposalTypesProps } from './types';
+
+export const testProps: ProposalTypesProps = {
+  data: [
+    {
+      title: 'Signal Proposal',
+      description:
+        'Lorem ipsum dolor sit amet, consectetur adipiscing elit. Sed leo quam, blandit eu sapien eu, commodo dapibus nisl',
+      onChainAction: false,
+    },
+    {
+      title: 'Custom Proposal',
+      description:
+        'Lorem ipsum dolor sit amet, consectetur adipiscing elit. Sed leo quam, blandit eu sapien eu, commodo dapibus nisl',
+      onChainAction: false,
+    },
+  ],
+};
