+++ src/Components/ProposalTypes/index.ts
@@ -1,0 +1,1 @@
+export { default as ProposalTypes } from './ProposalTypes';
