+++ src/Components/ProposalTypes/types.ts
@@ -1,0 +1,10 @@
+export interface ProposalTypeDescriptionProps {
+  title: string;
+  description: string;
+  icon?: React.FC<React.SVGProps<SVGSVGElement>>;
+  onChainAction: boolean;
+}
+
+export interface ProposalTypesProps {
+  data: ProposalTypeDescriptionProps[];
+}
